refactor: rename measurePerformance to measureRenders
