+++ README.md
@@ -94,11 +94,11 @@ Now that the library is installed, you can write your first test scenario in a f
 
 ```ts
 // ComponentUnderTest.perf-test.tsx
-import { measurePerformance } from 'reassure';
+import { measureRenders } from 'reassure';
 import { ComponentUnderTest } from './ComponentUnderTest';
 
 test('Simple test', async () => {
-  await measurePerformance(<ComponentUnderTest />);
+  await measureRenders(<ComponentUnderTest />);
 });
 ```
 
@@ -111,7 +111,7 @@ This test will measure render times of `ComponentUnderTest` during mounting and 
 If your component contains any async logic or you want to test some interaction, you should pass the `scenario` option:
 
 ```ts
-import { measurePerformance } from 'reassure';
+import { measureRenders } from 'reassure';
 import { screen, fireEvent } from '@testing-library/react-native';
 import { ComponentUnderTest } from './ComponentUnderTest';
 
@@ -121,7 +121,7 @@ test('Test with scenario', async () => {
     await screen.findByText('Done');
   };
 
-  await measurePerformance(<ComponentUnderTest />, { scenario });
+  await measureRenders(<ComponentUnderTest />, { scenario });
 });
 ```
 
@@ -130,7 +130,7 @@ The body of the `scenario` function is using familiar React Native Testing Libra
 In case of using a version of React Native Testing Library lower than v10.1.0, where [`screen` helper](https://callstack.github.io/react-native-testing-library/docs/api/#screen) is not available, the `scenario` function provides it as its first argument:
 
 ```ts
-import { measurePerformance } from 'reassure';
+import { measureRenders } from 'reassure';
 import { fireEvent } from '@testing-library/react-native';
 
 test('Test with scenario', async () => {
@@ -139,7 +139,7 @@ test('Test with scenario', async () => {
     await screen.findByText('Done');
   };
 
-  await measurePerformance(<ComponentUnderTest />, { scenario });
+  await measureRenders(<ComponentUnderTest />, { scenario });
 });
 ```
 
@@ -352,14 +352,14 @@ Looking at the example, you can notice that test scenarios can be assigned to ce
 
 ### Measurements
 
-#### `measurePerformance` function
+#### `measureRenders` function
 
 Custom wrapper for the RNTL `render` function responsible for rendering the passed screen inside a `React.Profiler` component,
 measuring its performance and writing results to the output file. You can use the optional `options` object that allows customizing aspects
 of the testing
 
 ```ts
-async function measurePerformance(
+async function measureRenders(
   ui: React.ReactElement,
   options?: MeasureOptions,
 ): Promise<MeasureResults> {
@@ -435,10 +435,11 @@ const defaultConfig: Config = {
 ```
 
 **`runs`**: the number of repeated runs in a series per test (allows for higher accuracy by aggregating more data). Should be handled with care.
+
 - **`warmupRuns`**: the number of additional warmup runs that will be done and discarded before the actual runs.
-**`outputFile`**: the name of the file the records will be saved to
-**`verbose`**: make Reassure log more, e.g. for debugging purposes
-**`testingLibrary`**: where to look for `render` and `cleanup` functions, supported values `'react-native'`, `'react'` or object providing custom `render` and `cleanup` functions
+  **`outputFile`**: the name of the file the records will be saved to
+  **`verbose`**: make Reassure log more, e.g. for debugging purposes
+  **`testingLibrary`**: where to look for `render` and `cleanup` functions, supported values `'react-native'`, `'react'` or object providing custom `render` and `cleanup` functions
 
 #### `configure` function
