+++ packages/reassure/src/index.ts
@@ -1,3 +1,3 @@
-export { measurePerformance, measureFunction, configure, resetToDefault } from '@callstack/reassure-measure';
+export { measureRenders, measureFunction, configure, resetToDefault } from '@callstack/reassure-measure';
 
 export { dangerReassure } from '@callstack/reassure-danger';
