+++ test-apps/native/src/SlowList.perf-test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { View, Text, Pressable } from 'react-native';
 import { fireEvent, screen } from '@testing-library/react-native';
-import { measurePerformance } from 'reassure';
+import { measureRenders } from 'reassure';
 import { SlowList } from './SlowList';
 
 const AsyncComponent = () => {
@@ -42,5 +42,5 @@ test('Async Component', async () => {
     await screen.findByText('Count: 5');
   };
 
-  await measurePerformance(<AsyncComponent />, { scenario });
+  await measureRenders(<AsyncComponent />, { scenario });
 });
